% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddBindingLifetime}
\alias{AddBindingLifetime}
\title{Add scanning characterization parameters}
\usage{
AddBindingLifetime(data.set, protein.filter, experimental.condition)
}
\arguments{
\item{data.set}{The output of the noise exclusion process.}
}
\value{
The input data set with several columns added for each parameter.
}
\description{
A function that calculates substrate binding lifetime,
scanning speed, scanning coverage and accumulative scanning range.
}
