% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddDurationFilter}
\alias{AddDurationFilter}
\title{Add trajectory duration filter}
\usage{
AddDurationFilter(data.set, minFrameDuration)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function.}

\item{minFrameDuration}{Minimum number of recorded frames for the signal to
be considered trajectory.}
}
\value{
The input data set with a added column where a binary label defines
whether a trajectory passes the duration filer or not.
}
\description{
A function that calculates the duration of trajectories and
labels them based on the given parameter.
}
