% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddFlankFilter}
\alias{AddFlankFilter}
\title{Add substrate associated label}
\usage{
AddFlankFilter(data.set, gauge)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function on which
AddDurationFilter() and AddVisualInspection() functions are applied.}

\item{guage}{An integer that defines how tightly the substrate-associated
filter is applied.}
}
\value{
The input data set with another column added that defines whether
a signal is substrate-associated or not.
}
\description{
A function that based on data from visually-inspected
trajectories creates a filtering system to label rest of the data.
}
