% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddIntensityFilter}
\alias{AddIntensityFilter}
\title{Add intensity range label}
\usage{
AddIntensityFilter(data.set, gauge)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function on which
AddDurationFilter() and AddVisualInspection() functions are applied.}

\item{guage}{An integer that defines how tightly the intensity range
filter is applied.}
}
\value{
The input data set with another column added that defines whether
the intensity of a signal is within the acceptable range or not.
}
\description{
A function that based on data from visually-inspected
trajectories creates an intensity filtering system that defines whether the
rest of the data is within the acceptable range or not.
}
