% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddRotatedCoord}
\alias{AddRotatedCoord}
\title{Extract the rotation slope and apply rotation matrix}
\usage{
AddRotatedCoord(data.set)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function on which
AddDurationFilter() and AddVisualInspection() functions are applied.}
}
\value{
The input data set where rotated coordinates are added for each data
point.
}
\description{
A parent function for extracting the rotation slope and applying
the rotation matrix for each data set.
}
