% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddStuckFilter}
\alias{AddStuckFilter}
\title{Add surface-associated label}
\usage{
AddStuckFilter(data.set, gauge)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function on which
AddDurationFilter() and AddVisualInspection() functions are applied.}

\item{guage}{An integer that defines how tightly the surface-associated
filter is applied.}
}
\value{
The input data set with another column added that defines whether
a signal is surface associated or not.
}
\description{
A function that based on data from surface-associated signals
creates a filtering system to label rest of the data.
}
