% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{AddVisualInspection}
\alias{AddVisualInspection}
\title{Add visual inspection indicator}
\usage{
AddVisualInspection(data.set, TrajsToInspect)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function.}

\item{TrajsToInspect}{Number of trajectories to inspect for each data set}
}
\value{
The input data set with a added column where a binary label defines
whether a trajectory is selected for visual inspection or not.
}
\description{
A function that picks longest ranging trajectories to send for
visusal inspection
}
