% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{ArrangeData}
\alias{ArrangeData}
\title{Arrange the detected trajectories}
\usage{
ArrangeData(RawDataSet, DetectedTrajectories)
}
\arguments{
\item{RawDataSet}{The output of the spatially filtered data set, this will
be used to join the additional attributes of each signal to the detected
trajectories.}

\item{DetectedTrajectories}{A long format of the output of the
FindTrajectory() function.}
}
\value{
A tidy data frame of all detected trajectories along with all the
signal localization attributes and metadata in separate columns.
}
\description{
A function that arranges all the detected trajectories according
to the frame number and gives trajectories unique IDs and  joins all the
attributes and metadata of the detected signal with the data at this point.
}
\details{
ArrangeData
}
