% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{ArrangeFilters}
\alias{ArrangeFilters}
\title{Define the type of the detected signal}
\usage{
ArrangeFilters(data.set)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function on which
AddDurationFilter(), AddVisualInspection(), AddStuckFilter(),
AddFlankFilter(), and AddIntensityFilter() functions are applied.}
}
\value{
The input data set with another column added that defines the type
of the detected signal based on information on existing filters.
}
\description{
A function based on the information from different filters
labels that type of the detected signal.
}
