% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{CorrectBlinking}
\alias{CorrectBlinking}
\title{Correct for discontinues trajectories}
\usage{
CorrectBlinking(InputData, x.distance, y.distance)
}
\arguments{
\item{InputData}{Long format of the data set including frame numbers and
x-y coordinates.}

\item{x.distance}{The maximum distance along the substrate that is allowed
for connecting points in trajectories.}

\item{y.distance}{The maximum distance across the substrate that is allowed
for connecting points in trajectories.}
}
\value{
A long format the data with frame number and x-y coordinates of
trajectories where blinking is corrected.
}
\description{
A function that correct for fluorescent dye blinking effect by
connecting segments of trajectories that are apart only within the defined
threshold.
}
\details{
CorrectBlinking
}
