% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{dependencies}
\alias{dependencies}
\title{Overall dependencies}
\description{
All the dependencies needed for the functions to run are listed below and
 are imported at once in this block:
magrittr, dplyr, tidyr, viridis, broom, data.table, DT, shiny, shinyWidgets,
shinyFiles, shinyscreenshot, stringr, scales, zoo, jsonlite, caret,
shinythemes, ggplot2, Rcpp, RcppRoll.
}
\details{
Dependencies
}
