% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{detectStuck}
\alias{detectStuck}
\title{Detect and arrange surface-associated signals}
\usage{
detectStuck(filteredDataSet, dxMax, dyMax)
}
\arguments{
\item{filteredDataSet}{The output of the spatially filtered data.}

\item{dxmax}{The maximum displacement of the signal along the x-direction,
before it is considered dissociated.this parameter is passed to the
FindTrajectory() function.}

\item{dymax}{The maximum displacement of the signal along the y-direction,
before it is considered dissociated. this parameter is passed to the
FindTrajectory() function.}
}
\value{
A tidy data frame of all detected surface-associated signal along
with all the signal localization attributes and metadata in separate columns
across all data sets.
}
\description{
A parent function for finding, arranging and collecting
the surface-associated signals across all existing data sets. The data
for each data set is run through a series other functions sequentially and
the end result is collected in tidy data frame.
}
