% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{FindTrajectory}
\alias{FindTrajectory}
\title{Finds trajectories of moving proteins}
\usage{
FindTrajectory(FilteredDataSet, dxmax, dymax)
}
\arguments{
\item{FilteredDataSet}{The data in the matrix representation where where unique
frames numbers are set as frame numbers and each signal on individual
columns.}

\item{dxmax}{The maximum displacement of the proteins along the substrate.}

\item{dymax}{The maximum displacement of the protein across the substrate
before it is considered dissociated.}
}
\value{
A matrix representation of the data where detected trajectories are
arranged in separate column and concatenated along the frame numbers.
}
\description{
A function that finds/traces the individual binding and scanning
events on the substrate. It rearranges the matrix representation of the
data so that continuous trajectories are set in columns of the matrix.
}
