% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{MakeLongForm}
\alias{MakeLongForm}
\title{Rearrange from the matrix form to tidy data}
\usage{
MakeLongForm(InputMatrix)
}
\arguments{
\item{InputDataSet}{A matrix representation of the data set.}
}
\value{
A four-column tidy representation of data with frame number and
x-y coordinates of trajectories along with their registered track.
}
\description{
A function that transforms the data from matrix format to
four-column tidy data.
}
\details{
MakeLongForm
}
