% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{MakeMatrix}
\alias{MakeMatrix}
\title{Create a matrix representation of the data frame}
\usage{
MakeMatrix(RawData)
}
\arguments{
\item{Rawdata}{Output of tranformData() function that is filtered for
each data set name.}
}
\value{
A matrix representation of the data frame where each row is indexed
as unique frame number and detected signals are spread in columns with x-y
coordinates.
}
\description{
A function that creates a matrix out of the provided data frame,
the row index is frame number and columns are number of simultaneously
localized signal and values are coordinates of the detected signal.
}
