% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{MakeTrajectoryAddress}
\alias{MakeTrajectoryAddress}
\title{Make trajectory address}
\usage{
MakeTrajectoryAddress(Input)
}
\arguments{
\item{Input}{The output of detectTrajectories() function on which
AddDurationFilter() and AddVisualInspection() functions are applied.}
}
\value{
A three-column data frame that contains file name, starting frame
number and ending frame number for each trajectory.
}
\description{
A function that extracts the temporal address of trajectories
that are selected for visual inspection.
}
