% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{MakeTrajectoryUniqeID}
\alias{MakeTrajectoryUniqeID}
\title{Make trajectory unique ID}
\usage{
MakeTrajectoryUniqeID(data.set)
}
\arguments{
\item{data.set}{The output of detectTrajectories() function.}
}
\value{
The input data set with a added column where each trajectories is
given a unique ID.
}
\description{
A function that labels trajectories with an unique ID and
arranges the rows based on data set name, trajectory unique ID and
frame number.
}
