% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{MakeZeroNA}
\alias{MakeZeroNA}
\title{Replace zero values with NA}
\usage{
MakeZeroNA(InputDataSet)
}
\arguments{
\item{InputDataSet}{A matrix representation of the data set.}
}
\value{
A matrix representation of the data set where zero values are replaced
by Na.
}
\description{
A function that replaces zero values in the data set with NA.
}
\details{
MakeZeroNA
}
