% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{readJsonFiles}
\alias{readJsonFiles}
\title{Read Json files, collect tabular data}
\usage{
readJsonFiles(dataPath, fileList, nEnd)
}
\arguments{
\item{dataPath}{A directory containing the json files}

\item{nEnd}{Number of characters to be included in the informative files name}
}
\value{
A structured tabular data set by combining all the existing data
}
\description{
A function that reads in the data from json files within a
 directory and collects all the existing data inside a tabular data frame.
}
