% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{readJsonFiles0}
\alias{readJsonFiles0}
\title{Read Json files, collect tabular data}
\usage{
readJsonFiles0(dataPath, nEnd)
}
\arguments{
\item{dataPath}{A directory containing the json files}

\item{nEnd}{Number of characters to be included in the informative files name}

\item{fileList}{A list of json files inside the directory}
}
\value{
A structured tabular data set by combining all the existing data
}
\description{
A function that reads in the data from json files within a
 directory and collects all the existing data inside a tabular data frame.
}
